Complete part 4 exercises up to 4.22

+++ part4/blog/controllers/blogs.js
@@ -1,27 +1,67 @@
 
+require('dotenv').config();
+const jwt = require('jsonwebtoken');
+
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
+const User = require('../models/user');
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1, id: 1 });
   response.json(blogs);
 });
 
 blogsRouter.post('/', async (request, response) => {
   const blog = request.body;
 
+  const decodedToken = jwt.verify(blog.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid or missing' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
   if (!blog.likes) {
     blog.likes = 0;
   }
 
-  const newBlog = new Blog(request.body);
+  const newBlog = new Blog({
+    ...blog,
+    // eslint-disable-next-line no-underscore-dangle
+    user: user._id,
+  });
+
   const savedBlog = await newBlog.save();
-  response.status(201).json(savedBlog);
+
+  // eslint-disable-next-line no-underscore-dangle
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  return response.status(201).json(savedBlog);
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const { body } = request;
+
+  const decodedToken = jwt.verify(body.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid or missing' });
+  }
+
+  const user = await User.findById(decodedToken.id);
+
+  const blog = await Blog.findById(request.params.id);
+
+  // eslint-disable-next-line no-underscore-dangle
+  if (user._id.toString() !== blog.user.toString()) {
+    return response.status(401)
+      .json({ error: 'user not authorized to delete the blog' });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  return response.status(204).end();
 });
 
 blogsRouter.put('/:id', async (request, response) => {
